+++ src/pipeline_telemetry/__init__.py
@@ -32,6 +32,8 @@ Arguments
 
     -
 """
+from .main import Telemetry, mongo_telemetry  # noqa: F401
 from .settings import errors  # noqa: F401
 from .settings import process_type  # noqa: F401
+from .settings.process_type import ProcessTypes  # noqa: F401
 from .validators import has_key, validate_entries  # noqa: F401