+++ tests/test_telemetry.py
@@ -6,11 +6,12 @@ from datetime import datetime
 import pytest
 from test_data import DEFAULT_TELEMETRY_PARAMS
 
-from pipeline_telemetry.main import FAIL_COUNT, Telemetry
+from pipeline_telemetry.main import FAIL_COUNT, Telemetry, mongo_telemetry
 from pipeline_telemetry.settings import exceptions, settings
 from pipeline_telemetry.settings.settings import ProcessType
-from pipeline_telemetry.storage import AbstractTelemetryStorage, \
-    TelemetryInMemoryStorage
+from pipeline_telemetry.storage.generic import AbstractTelemetryStorage
+from pipeline_telemetry.storage.memory import TelemetryInMemoryStorage
+from pipeline_telemetry.storage.mongo import TelemetryMongoStorage
 
 # pylint: disable=protected-access
 
@@ -32,7 +33,7 @@ def test_telemetry_instance_creation_raises_excption():
     ProcessType.
     """
     process_type = ProcessType(
-        process_name='not_registered',
+        process_type='not_registered',
         subtypes=['test'])
     telemetry_params = {
         'process_name': 'load_weather_data',
@@ -263,12 +264,13 @@ def test_close_telemetry_instance_sets_run_time(mocker):
     check that closing the telemetry sets the run_time
     """
     mocker.patch(
-        'pipeline_telemetry.storage.TelemetryInMemoryStorage.store_telemetry',
+        ("pipeline_telemetry.storage.memory."
+         "TelemetryInMemoryStorage.store_telemetry"),
         return_value=None)
     telemetry_inst = Telemetry(**DEFAULT_TELEMETRY_PARAMS)
     telemetry_inst.increase_sub_process_base_count('RETRIEVE_RAW_DATA')
     telemetry_result = telemetry_inst.save_and_close()
-    assert telemetry_result.get('run_time_in_seconds') > 0
+    assert float(telemetry_result.get('run_time_in_seconds')) > 0
 
 
 def test_close_telemetry_instance_calls_store_telemetry(mocker):
@@ -276,7 +278,8 @@ def test_close_telemetry_instance_calls_store_telemetry(mocker):
     check that closing the telemetry sets the run_time
     """
     mocker.patch(
-        'pipeline_telemetry.storage.TelemetryInMemoryStorage.store_telemetry',
+        ("pipeline_telemetry.storage.memory."
+         "TelemetryInMemoryStorage.store_telemetry"),
         return_value=None)
     _store_telemetry_spy = mocker.spy(
         TelemetryInMemoryStorage, 'store_telemetry')
@@ -332,3 +335,13 @@ def test_storage_class_close_method_closes_db():
     storage_instance.close_db()
     assert storage_instance.db_in_memory is None
     assert storage_instance.db_cursor is None
+
+
+def test_mongo_telemetry():
+    """
+    Test mongo_telemetry method returns a Telemetry object with a mongo storage
+    class.
+    """
+    telemetry = mongo_telemetry(telemetry_rules={}, **DEFAULT_TELEMETRY_PARAMS)
+    assert isinstance(telemetry, Telemetry)
+    assert isinstance(telemetry._storage_class, TelemetryMongoStorage)