+++ CHANGELOG.rst
@@ -6,3 +6,12 @@ Changelog
 ------------------
 
 * First release on PyPI.
+
+
+0.0.4 (2021-10-31)
+------------------
+
+* Seperated the storage module
+* Added MongoDb storage class
+* Started with documentation
+